requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0

# AI/ML
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3

# Data Processing
pandas==2.2.0

# API & HTTP
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1

# Caching
redis==5.0.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# Image Processing
opencv-python==4.9.0.80

# Search & Text Processing
scikit-learn==1.4.0
sentence-transformers==2.3.1

# Development
pytest==7.4.4
black==24.1.1
flake8==7.0.0
